+++ fj-doc-sample/src/test/java/test/org/fugerit/java/doc/sample/dev/TestFo01.java
@@ -16,4 +16,5 @@ class TestFo01 extends DevHelper {
 				new File( BasicFacadeTest.BASIC_OUTPUT_PATH, testCase+".pdf" ) );
 		Assertions.assertTrue( res );
 	}
+
 }