+++ CHANGELOG.md
@@ -10,10 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Image validator based on TwelveMonkeys ImageIO (currently supporting TIFF) <https://github.com/fugerit-org/fj-doc/issues/551>
+- fj-doc-mod-fop, new mod-fop-xslt-path property <https://github.com/fugerit-org/fj-doc/issues/558>
 
 ### Changed
 
 - quarkus-version set to 3.26.5 across all the modules
+
+### Fixed
+
+- DocInput.getDoc() now reuse docBase
 
 ## [8.17.0] - 2025-10-14